Basic Operations have implemented.

+++ src/main/java/com/example/banking/data/repository/PrimaryAccountRepository.java
@@ -6,6 +6,6 @@ import com.example.banking.data.entity.PrimaryAccount;
 
 public interface PrimaryAccountRepository extends CrudRepository<PrimaryAccount, Long> {
 
-	public PrimaryAccount findByAccountNumber(int accountNumber);
+	public PrimaryAccount findByAccountNumber(long accountNumber);
 
 }
